Fix comments and get ready to merge
