+++ src/eladkay/ellipticcurve/mathengine/EllipticCurveHelper.kt
@@ -1,5 +1,8 @@
 package eladkay.ellipticcurve.mathengine
 
+import eladkay.ellipticcurve.simulationengine.CurveFrame
+import eladkay.ellipticcurve.simulationengine.EllipticSimulator
+
 class EllipticCurveHelper(private val curve: EllipticCurve) {
     private val field: Field = curve.field
 
@@ -34,5 +37,7 @@ class EllipticCurveHelper(private val curve: EllipticCurve) {
         else
             add(a, multiply(a, num - 1))
     }
+
+
 
 }
