+++ src/eladkay/ellipticcurve/mathengine/Vec2i.kt
