+++ src/eladkay/ellipticcurve/simulationengine/CurveFrame.kt
@@ -1,6 +1,7 @@
 package eladkay.ellipticcurve.simulationengine
 
 import eladkay.ellipticcurve.mathengine.EllipticCurve
+import eladkay.ellipticcurve.mathengine.Field
 import eladkay.ellipticcurve.mathengine.Vec2i
 import java.awt.Color
 
@@ -17,4 +18,24 @@ interface CurveFrame {
     fun clear()
     var curve: EllipticCurve
     operator fun <T> invoke(action: CurveFrame.() -> T) = action()
+
+    fun serializeCurveFrame(): String {
+        val curve = curve
+        val scale = EllipticSimulator.scale
+        val field = if(curve.field == Field.REALS) "R" else (curve.field as Field.ZpField).modulo.toString()
+        return "${curve.aValue};${curve.bValue};$field;$scale"
+    }
+
+    companion object {
+
+        fun deserializeCurveFrame(string: String): EllipticCurve {
+            val split = string.split(";")
+            val field = if(split[2] == "R") Field.REALS else Field.createModuloField(split[2].toInt())
+            val curve = EllipticCurve(split[0].toDouble(), split[1].toDouble(), field)
+            EllipticSimulator.scale = split[3].toInt()
+            return curve
+        }
+    }
+
+
 }
